+++ src/components/blockRenderer/specialBlocks/slider.js
@@ -1,9 +1,11 @@
 import { Component, createRef } from 'react'
 import equal from 'react-fast-compare'
 
-import _b5BlocksObject from '../../../b5.js/src/blocks/blocksObjectWrapper'
 import Node from '../node'
 import { InputBox } from '../frags'
+
+import ThumbArrow from '../../../img/components/arrow.svg'
+import { getOutputConnectType } from '../blockRendererMethod'
 
 function constrain(v, a, b) {
   // Constrain v within a and b without
@@ -14,80 +16,50 @@ function constrain(v, a, b) {
   return max(min(v, max(a, b)), min(a, b))
 }
 
-export class InputBlock extends Component {
-  render() {
-    const {
-      action,
-      className,
-      name,
-      text,
-      type,
-      x,
-      y,
-      inlineData,
-      output,
-      outputNodes,
-      collect,
-      nodesRef,
-      focused,
-      selectedNodes,
-    } = this.props
-    return (
-      // * string input block is specially long...
-      <div
-        className={className + (text === 'str' ? ' longInputWidth' : '')}
-        data-hints={`${name} block`}
-      >
-        {/* <div className="blockName">{text}</div> */}
-        <InputBox
-          action={action}
-          className={''}
-          thisInlineData={inlineData[0]}
-          thisDataType={_b5BlocksObject.original[name].inlineData[0].type[1]}
-          inlineDataInd={0}
-          name={name}
-          x={x}
-          y={y}
-          collect={collect}
-          refPosition={0}
-        />
-        <p className="nodeText bottomText">{text}</p>
-        {output && (
-          <div className="nodes outputNodes">
-            <Node
-              nodeClass="output"
-              count={1}
-              type={type}
-              connectType={
-                output[0].length !== 0 ? outputNodes[0].type[0] : null
-              }
-              ref={nodesRef.output[0]}
-              focused={focused}
-              selected={selectedNodes.output.includes('0')}
-              name={name}
-              nodeType={'outputNodes'}
-              refPosition={0}
-              hintSide={'down'}
-            />
-          </div>
-        )}
-      </div>
-    )
-  }
+const countDecimalPlaces = n => {
+  if (Math.floor(n.valueOf()) === n.valueOf()) return 0
+  return n.toString().split('.')[1].length || 0
 }
 
 const _inputRangeThumbWidth = 1
 const _stepRange = [0, Infinity]
 
+const _getLength = (totalLen, current, min, max) => {
+  return current === min
+    ? 0
+    : current === max
+    ? totalLen
+    : totalLen * ((current - min) / (max - min))
+}
+
+const _getValue = (totalLen, length, min, max, step) => {
+  // Get value from slider length
+  return length === 0
+    ? min
+    : length === totalLen
+    ? max
+    : Math.max(
+        Math.min(
+          Math.round(((length / totalLen) * (max - min) + min) / step) * step,
+          max
+        ),
+        min
+      )
+}
+
 class InputRange extends Component {
   constructor(props) {
     super(props)
     this.state = { currentValue: props.inlineData[0] }
+
+    this.textBoxPosition = 'r'
+    this.currentInput = createRef()
   }
 
   componentDidMount() {
     this.totalLength = this.sliderBox.offsetWidth - _inputRangeThumbWidth
     this._setPosition()
+    this._placeTextBox()
   }
 
   componentDidUpdate() {
@@ -115,39 +87,42 @@ class InputRange extends Component {
       collect([x, y, 0, newCurrent], 'inlineDataChange')
       this.setState({ currentValue: newCurrent })
     }
-    this.rangeBelow.style.width = this._getLength(newCurrent, min, max) + 'px'
-  }
-
-  _getLength = (current, min, max) => {
-    // Get length from current slider value
-    return current === min
-      ? 0
-      : current === max
-      ? this.totalLength
-      : this.totalLength * ((current - min) / (max - min))
-  }
-
-  _getValue = (length, min, max, step) => {
-    // Get value from slider length
-    return length === 0
-      ? min
-      : length === this.totalLength
-      ? max
-      : Math.max(
-          Math.min(
-            Math.round(
-              ((length / this.totalLength) * (max - min) + min) / step
-            ) * step,
-            max
-          ),
-          min
-        )
+    this.rangeBelow.style.width =
+      _getLength(this.totalLength, newCurrent, min, max) + 'px'
   }
 
   collectRangeData() {
     // Collect current data to editor
     const { collect, x, y } = this.props
     collect([x, y, 0, this.state.currentValue], 'inlineDataChange')
+  }
+
+  _placeTextBox = () => {
+    if (this.rangeBelow.offsetWidth > this.totalLength * 0.65) {
+      if (this.textBoxPosition === 'r') {
+        this.textBoxPosition = 'l'
+        this.currentInput.current.style.left = 'unset'
+        this.currentInput.current.style.right = '15px' // calc($thumb-width + $gap)
+      }
+    } else {
+      if (this.textBoxPosition === 'l') {
+        this.textBoxPosition = 'r'
+        this.currentInput.current.style.left = '15px'
+        this.currentInput.current.style.right = 'unset'
+      }
+    }
+
+    // Arrows
+    if (this.rangeBelow.offsetWidth > this.totalLength - 8) {
+      this.rArrow.style.display = 'none'
+    } else if (this.rangeBelow.offsetWidth < 8) {
+      this.lArrow.style.display = 'none'
+    } else if (
+      this.lArrow.style.display !== 'block' ||
+      this.rArrow.style.display !== 'block'
+    ) {
+      this.lArrow.style.display = this.rArrow.style.display = 'block'
+    }
   }
 
   handleSlide = e => {
@@ -164,7 +139,8 @@ class InputRange extends Component {
       const handleDrag = e => {
         let deltaX = (e.clientX - mouse.x) / this.props.scale
 
-        let value = this._getValue(
+        let value = _getValue(
+          this.totalLength,
           Math.max(Math.min(deltaX + mouse.width, this.totalLength), 0),
           iD[1],
           iD[2],
@@ -174,7 +150,11 @@ class InputRange extends Component {
           this.setState({ currentValue: value })
 
         this.rangeBelow.style.width =
-          this._getLength(value, iD[1], iD[2]) + 'px'
+          _getLength(this.totalLength, value, iD[1], iD[2]) + 'px'
+
+        // Move current text box to the left or right if needed
+        this._placeTextBox()
+
         this.collectRangeData()
       }
 
@@ -196,7 +176,7 @@ class InputRange extends Component {
   }
 
   render() {
-    const { action, name, x, y, inlineData, collect } = this.props
+    const { action, name, x, y, inlineData, collect, fraction } = this.props
     const { currentValue } = this.state
 
     return (
@@ -211,41 +191,62 @@ class InputRange extends Component {
           ref={e => (this.rangeBelow = e)}
           className="rangeBelow sliderComponent"
         ></div>
+        {/* Thumb */}
         <div
           ref={e => (this.thumb = e)}
           onMouseDown={action ? this.handleSlide : null}
           className="thumb sliderComponent ewResizing"
         >
+          <img
+            ref={e => (this.lArrow = e)}
+            src={ThumbArrow}
+            className="thumbArrowLeft thumbArrow"
+            alt=""
+          />
+          <img
+            ref={e => (this.rArrow = e)}
+            src={ThumbArrow}
+            className="thumbArrowRight thumbArrow"
+            alt=""
+          />
+          {/* CURRENT */}
           <InputBox
+            ref={this.currentInput}
             action={action}
             className={' current'}
-            thisInlineData={currentValue}
+            thisInlineData={currentValue.toFixed(
+              countDecimalPlaces(inlineData[3])
+            )}
             thisDataType={'number'}
             inlineDataInd={0}
             name={name}
             x={x}
             y={y}
             collect={collect}
-            refPosition={0}
+            hintRefPosition={0}
           />
         </div>
 
         <div ref={e => (this.sliderInput = e)} className="sliderInput">
-          <InputBox
-            action={action}
-            className={' sliderSet min'}
-            thisInlineData={inlineData[1]}
-            thisDataType={'number'}
-            inlineDataInd={1}
-            name={name}
-            x={x}
-            y={y}
-            collect={collect}
-            refPosition={1}
-            hintSide={'down'}
-          />
+          {/* MIN */}
+          {!fraction && (
+            <InputBox
+              action={action}
+              className={' sliderSet min'}
+              thisInlineData={inlineData[1]}
+              thisDataType={'number'}
+              inlineDataInd={1}
+              name={name}
+              x={x}
+              y={y}
+              collect={collect}
+              hintRefPosition={1}
+              hintSide={'down'}
+            />
+          )}
 
           <p className="nodeText bottomText stepText">step</p>
+          {/* STEP */}
           <InputBox
             action={action}
             className={' sliderSet step'}
@@ -257,30 +258,40 @@ class InputRange extends Component {
             x={x}
             y={y}
             collect={collect}
-            refPosition={3}
+            hintRefPosition={3}
             hintSide={'down'}
           />
 
-          <InputBox
-            action={action}
-            className={' sliderSet max'}
-            thisInlineData={inlineData[2]}
-            thisDataType={'number'}
-            inlineDataInd={2}
-            name={name}
-            x={x}
-            y={y}
-            collect={collect}
-            refPosition={2}
-            hintSide={'down'}
-          />
+          {/* MAX */}
+          {!fraction && (
+            <InputBox
+              action={action}
+              className={' sliderSet max'}
+              thisInlineData={inlineData[2]}
+              thisDataType={'number'}
+              inlineDataInd={2}
+              name={name}
+              x={x}
+              y={y}
+              collect={collect}
+              hintRefPosition={2}
+              hintSide={'down'}
+            />
+          )}
         </div>
       </div>
     )
   }
 }
 
-export class SliderBlock extends Component {
+class Slider extends Component {
+  constructor(props) {
+    super(props)
+    this.sliderRef = createRef()
+  }
+}
+
+export default class SliderBlock extends Slider {
   /*
   
   <SliderBlock>
@@ -290,11 +301,6 @@ export class SliderBlock extends Component {
   </SliderBlock>
 
   */
-  constructor(props) {
-    super(props)
-    this.sliderRef = createRef()
-  }
-
   render() {
     const {
       action,
@@ -305,6 +311,8 @@ export class SliderBlock extends Component {
       x,
       y,
       inlineData,
+      input,
+      inputNodes,
       output,
       outputNodes,
       collect,
@@ -312,11 +320,30 @@ export class SliderBlock extends Component {
       focused,
       selectedNodes,
       scale,
+      fraction,
     } = this.props
 
     return (
       <div className={className} data-hints={`${name} block`}>
-        {/* <div className="blockName">{text}</div> */}
+        {fraction && (
+          <div className="nodes inputNodes">
+            <Node
+              nodeClass="input"
+              count={1}
+              type={type}
+              connectType={input[0] !== null ? inputNodes[0].type[0] : null}
+              ref={nodesRef.input[0]}
+              focused={focused}
+              selected={selectedNodes.input.includes('0')}
+              name={name}
+              nodeType={'inputNodes'}
+              hintRefPosition={0}
+              hintSide={'up'}
+              background={true}
+            />
+          </div>
+        )}
+
         <InputRange
           action={action}
           name={name}
@@ -325,21 +352,26 @@ export class SliderBlock extends Component {
           x={x}
           y={y}
           scale={scale}
+          fraction={fraction}
         />
 
-        <p className="nodeText bottomText">{text}</p>
+        <p
+          className={'nodeText bottomText' + (fraction ? ' fractionText' : '')}
+        >
+          {text}
+        </p>
         <div className="nodes outputNodes">
           <Node
             nodeClass="output"
             count={1}
             type={type}
-            connectType={output[0].length !== 0 ? outputNodes[0].type[0] : null}
+            connectType={getOutputConnectType(output, outputNodes, 0)}
             ref={nodesRef.output[0]}
             focused={focused}
             selected={selectedNodes.output.includes('0')}
             name={name}
             nodeType={'outputNodes'}
-            refPosition={0}
+            hintRefPosition={0}
             hintSide={'down'}
           />
         </div>