+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- New blocks including transformation, etc.
+- New blocks including transformation, fraction slider, etc.
 - Upgrade to React v17.0.
 
 ### Changed
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A much lighter and faster way to parse hint for components.
 - Migrate from p5 to [q5](https://github.com/peilingjiang-DEV/q5xjs.git) for better performance.
 - Move (pan) around the canvas by all three keys (left, middle, right) of the mouse, instead of only the right one.
+- New look of the sliders.
 
 ### Removed
 
@@ -27,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Code canvas stops taking actions after adding a new block.
 - Errors in the following blocks: `rotate`.
 - Canvas `wheel` event can't prevent default.
+- "Add block" hint doesn't appear when no block is in the canvas.
 
 ## [0.1.2] - 2020.12.23